Fix drawing lines between datatime data points

+++ src/backend/worksheet/plots/cartesian/XYCurvePrivate.h
@@ -47,9 +47,9 @@ public:
 	void retransform();
 	void recalcLogicalPoints();
 	void updateLines();
-	void addLine(QPointF p0, QPointF p1, QPointF& lastPoint, int& pixelDiff, int numberOfPixelX); // for any x scale
-	void addLinearLine(QPointF p0, QPointF p1, QPointF& lastPoint, double minLogicalDiffX, int& pixelDiff);	// optimized for linear x scale
-	void addUniqueLine(QPointF p0, QPointF p1, QPointF& lastPoint, int& pixelDiff);	// finally add line if unique (no overlay)
+	void addLine(QPointF p0, QPointF p1, QPointF& lastPoint, qint64& pixelDiff, int numberOfPixelX); // for any x scale
+	void addLinearLine(QPointF p0, QPointF p1, QPointF& lastPoint, double minLogicalDiffX, qint64& pixelDiff);	// optimized for linear x scale
+	void addUniqueLine(QPointF p0, QPointF p1, QPointF& lastPoint, qint64& pixelDiff);	// finally add line if unique (no overlay)
 	void updateDropLines();
 	void updateSymbols();
 	void updateValues();
